requirements.txt
pyinfra
lastversion
pytest
pytest-testinfra
python-dotenv